Change line height to 1.3
This is the smallest value that I could find that doesn't cause
scrollbars to be displayed on `<pre>` elements.  The problem there is
that the line height is too small to allow the box to be sized
correctly; the ascender or descender height of the font cause the actual
line height to exceed this value.  Setting a larger line height avoids
the problem, at a cost of extra vertical space being used.

The alternative is to remove the overflow attribute, but that would
cause overlarge figures, which are more common than they should be, to
be rendered poorly.

cc @larseggert for awareness